Add small & secondaries explosions, refine fire
Introduce two new explosion types ("Small explosion" and "Secondaries") across frontend and backend, add handling in maps/menus, and tweak fire behavior. Change Explosion.intensity from unsigned int to double and reduce AI grenade spawn intensity to 0.1. Frontend: expose new options in effect/unit explosion menus and map handlers to spawn the appropriate server effects. Lua backend: rename 'secondary' -> 'secondaries', make createFire accept preset and density (used by Olympus.explosion and random debris), improve secondaries to spawn delayed debris explosions and spawn smaller fires. These changes enable more granular explosion intensities and a multi-stage 'secondaries' effect.

+++ backend/core/include/commands.h
@@ -428,7 +428,7 @@ private:
 class Explosion : public Command
 {
 public:
-	Explosion(unsigned int intensity, string explosionType, Coords location, function<void(void)> callback = [](){}) :
+	Explosion(double intensity, string explosionType, Coords location, function<void(void)> callback = [](){}) :
 		Command(callback),
 		location(location),
 		intensity(intensity),
@@ -441,7 +441,7 @@ public:
 
 private:
 	const Coords location;
-	const unsigned int intensity;
+	const double intensity;
 	const string explosionType;
 };
